+++ src/Services/AbstractService.php
@@ -4,7 +4,10 @@ namespace Bigbank\DigiDoc\Services;
 use Bigbank\DigiDoc\Soap\DigiDocServiceInterface;
 
 /**
- * {@inheritdoc}
+ * Common base class for DigiDoc service classes
+ *
+ * `DigiDocServiceInterface` combines all publicly available SOAP
+ * methods and is needed by every service.
  */
 abstract class AbstractService
 {