+++ src/Services/MobileId/Authenticator.php
@@ -3,7 +3,9 @@ namespace Bigbank\DigiDoc\Services\MobileId;
 
 use Bigbank\DigiDoc\Exceptions\DigiDocException;
 use Bigbank\DigiDoc\Services\AbstractService;
+use Bigbank\DigiDoc\Soap\DigiDocServiceInterface;
 use Bigbank\DigiDoc\Soap\InteractionStatus;
+use RandomLib\Generator;
 
 /**
  * {@inheritdoc}
@@ -11,12 +13,37 @@ use Bigbank\DigiDoc\Soap\InteractionStatus;
 class Authenticator extends AbstractService implements AuthenticatorInterface
 {
 
+    /**
+     * Time to wait (in seconds) before starting polling for authentication status
+     *
+     * Used by `waitForAuthentication`, but not by `askStatus`
+     */
+    const INITIAL_WAIT_PERIOD = 8;
+
+    /**
+     * The length of client-side random token 'SPChallenge'
+     */
     const SP_CHALLENGE_LENGTH = 20;
 
     /**
-     * @var int
+     * @var int Amount of time in seconds between poll requests for the `waitForAuthentication` query
      */
     protected $pollingFrequency = 3;
+
+    /**
+     * @var Generator
+     */
+    private $random;
+
+    /**
+     * @param DigiDocServiceInterface $digiDocService
+     * @param Generator               $random
+     */
+    public function __construct(DigiDocServiceInterface $digiDocService, Generator $random)
+    {
+        parent::__construct($digiDocService);
+        $this->random = $random;
+    }
 
     /**
      * {@inheritdoc}
@@ -48,7 +75,7 @@ class Authenticator extends AbstractService implements AuthenticatorInterface
         $statusResponse = $this->digiDocService->GetMobileAuthenticateStatus($sessionCode, false);
 
         if (!isset($statusResponse['Status'])) {
-            throw new DigiDocException;
+            throw new DigiDocException('DigiDoc response does not include all required elements');
         }
         return $statusResponse['Status'];
     }
@@ -58,6 +85,11 @@ class Authenticator extends AbstractService implements AuthenticatorInterface
      */
     public function waitForAuthentication($sessionCode, callable $callback)
     {
+        // Initial sleep period
+        // It is reasonable to wait before starting sending status queries - it is
+        // improbable that message from user’s phone arrives earlier because of technical and
+        // human limitations.
+        sleep(self::INITIAL_WAIT_PERIOD);
 
         $status = InteractionStatus::OUTSTANDING_TRANSACTION;
         while ($status == InteractionStatus::OUTSTANDING_TRANSACTION) {
@@ -69,29 +101,14 @@ class Authenticator extends AbstractService implements AuthenticatorInterface
     }
 
     /**
-     * Generates a random string for SP_CHALLENGE parameter of DigiDoc
+     * Generates a random hexadecimal string for SPChallenge parameter of DigiDoc
      *
-     * The generated string is cryptographically secure if the function `random_bytes` is available (>= PHP 7).
+     * The generated string is cryptographically secure.
      *
      * @return string
      */
     private function generateChallenge()
     {
-
-        return substr($this->generateRandomString(), 0, self::SP_CHALLENGE_LENGTH);
-    }
-
-    /**
-     * Return a 40-char random string
-     *
-     * @return string
-     */
-    private function generateRandomString()
-    {
-
-        if (function_exists('random_bytes')) {
-            return bin2hex(random_bytes(self::SP_CHALLENGE_LENGTH));
-        }
-        return bin2hex(openssl_random_pseudo_bytes(self::SP_CHALLENGE_LENGTH));
+        return $this->random->generateString(self::SP_CHALLENGE_LENGTH, '0123456789ABCDEF');
     }
 }