+++ tests/Services/MobileId/AuthenticatorTest.php
@@ -4,7 +4,10 @@ namespace Bigbank\DigiDoc\Test;
 
 use Bigbank\DigiDoc\Services\MobileId\Authenticator;
 use Bigbank\DigiDoc\Soap\DigiDocServiceInterface;
+use Bigbank\DigiDoc\Soap\InteractionStatus;
 use Bigbank\DigiDoc\Test\Soap\DummyDigiDocService;
+use RandomLib\Factory;
+use RandomLib\Generator;
 
 /**
  * @coversDefaultClass \Bigbank\DigiDoc\Services\MobileId\Authenticator
@@ -13,17 +16,34 @@ class AuthenticatorTest extends TestCase
 {
 
     /**
+     * @var \PHPUnit_Framework_MockObject_MockObject|DigiDocServiceInterface $digiDocMock
+     */
+    protected $digiDocMock;
+
+    /**
+     * @var Generator
+     */
+    protected $random;
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function setUp()
+    {
+        parent::setUp();
+        $this->digiDocMock = $this->getMock(
+            DummyDigiDocService::class,
+            ['MobileAuthenticate', 'GetMobileAuthenticateStatus']
+        );
+        $this->random      = (new Factory)->getLowStrengthGenerator();
+    }
+
+    /**
      * @covers ::generateChallenge
      */
-    public function testGenerateChallengeReturnsUtf8StringWhenRunWithPhp7()
+    public function testGenerateChallengeReturnsHexStringOfCorrectLength()
     {
-
-        $this->requirePhpVersion('7.0');
-
-        /** @var \PHPUnit_Framework_MockObject_MockObject|DigiDocServiceInterface $digiDocMock */
-        $digiDocMock = $this->getMock(DummyDigiDocService::class, ['MobileAuthenticate']);
-
-        $digiDocMock->expects($this->once())
+        $this->digiDocMock->expects($this->once())
             ->method('MobileAuthenticate')
             ->with(
                 '14212128025',
@@ -33,14 +53,113 @@ class AuthenticatorTest extends TestCase
                 null,
                 null,
                 $this->callback(function ($randomString) {
-                    return mb_check_encoding($randomString, 'UTF-8');
+                    // Verify that generated challenge is only of HEX characters and of fixed length
+                    return ctype_xdigit($randomString)
+                    && mb_strlen($randomString) === Authenticator::SP_CHALLENGE_LENGTH;
                 }),
                 'asynchClientServer',
                 null,
                 false,
                 false
             );
-        $authenticator = new Authenticator($digiDocMock);
+        $authenticator = $this->authenticatorFactory();
         $authenticator->authenticate('14212128025', '37200007', null, null);
+    }
+
+
+    /**
+     * @covers ::waitForAuthentication
+     */
+    public function testWaitForAuthenticationContinuesWhenStatusChanges()
+    {
+        $authenticator = $this->getMock(Authenticator::class, ['askStatus'], [$this->digiDocMock, $this->random]);
+        $authenticator->expects($this->exactly(2))
+            ->method('askStatus')
+            ->will($this->onConsecutiveCalls(
+                InteractionStatus::OUTSTANDING_TRANSACTION,
+                InteractionStatus::USER_CANCEL
+            ));
+
+        $authenticator->waitForAuthentication(
+            0,
+            function () {
+            }
+        );
+    }
+
+    /**
+     * @covers ::waitForAuthentication
+     */
+    public function testWaitForAuthenticationCallsCallbackFunction()
+    {
+        $authenticator = $this->getMock(Authenticator::class, ['askStatus'], [$this->digiDocMock, $this->random]);
+        $authenticator->expects($this->once())
+            ->method('askStatus')
+            ->willReturn(InteractionStatus::USER_CANCEL);
+
+        $callback = $this->getMock(\stdClass::class, ['callback']);
+        $callback->expects($this->once())
+            ->method('callback')
+            ->with(InteractionStatus::USER_CANCEL, 444);
+
+        $authenticator->waitForAuthentication(444, [$callback, 'callback']);
+    }
+
+    /**
+     * @covers ::waitForAuthentication
+     */
+    public function testWaitForAuthenticationHasAnInitialWaitPeriodBeforePolling()
+    {
+
+        $start = microtime(true);
+
+        $authenticator = $this->getMock(Authenticator::class, ['askStatus'], [$this->digiDocMock, $this->random]);
+        $authenticator->expects($this->once())
+            ->method('askStatus')
+            ->with($this->callback(function () use ($start) {
+                // Verify that there's a sleep call before the first call to askStatus
+                return microtime(true) - $start > Authenticator::INITIAL_WAIT_PERIOD;
+            }))
+            ->willReturn(InteractionStatus::USER_CANCEL);
+
+        $authenticator->waitForAuthentication(
+            0,
+            function () {
+            }
+        );
+    }
+
+    /**
+     * @covers ::askStatus
+     * @expectedException \Bigbank\DigiDoc\Exceptions\DigiDocException
+     */
+    public function testAskStatusThrowsOnInvalidResponse()
+    {
+        $this->digiDocMock->expects($this->once())
+            ->method('GetMobileAuthenticateStatus')
+            ->willReturn([]);
+
+        $this->authenticatorFactory()->askStatus(0);
+    }
+
+    /**
+     * @covers ::askStatus
+     */
+    public function testAskStatusReturnsStatusAsString()
+    {
+        $this->digiDocMock->expects($this->once())
+            ->method('GetMobileAuthenticateStatus')
+            ->willReturn(['Status' => InteractionStatus::USER_CANCEL]);
+
+        $status = $this->authenticatorFactory()->askStatus(0);
+        $this->assertSame(InteractionStatus::USER_CANCEL, $status);
+    }
+
+    /**
+     * @return Authenticator
+     */
+    private function authenticatorFactory()
+    {
+        return new Authenticator($this->digiDocMock, $this->random);
     }
 }